Signin / out button workign
