+++ components/Navigation.tsx
@@ -5,9 +5,7 @@ import Image from "next/image";
 import { usePathname } from "next/navigation";
 import clsx from "clsx";
 
-export default function Navigation({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
+export default function Navigation() {
   const [isOpen, setIsOpen] = useState(false);
   const pathname = usePathname();
 
@@ -22,10 +20,7 @@ export default function Navigation({
   };
 
   return (
-    <div
-      className="flex mb-24 justify-between py-6 px-10 relative mr-7 z-10 items-start"
-      data-testid="navbar"
-    >
+    <>
       <Link href="/" className="text-lg mr-10 2lg:mr-16 ">
         <Image src="/logo.png" alt="Fruit Finder Logo" width="50" height="50" />
       </Link>
@@ -80,7 +75,6 @@ export default function Navigation({
           </li>
         ))}
       </ul>
-      <div className="hidden lg:flex items-center">{children}</div>
-    </div>
+    </>
   );
 }
